aktc/requirements.txt
asgiref==3.8.1
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.3
django-appconf==1.1.0
django-colorfield==0.14.0
django-compressor==4.5.1
django-cors-headers==4.7.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-slugify==8.0.4
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
uvicorn==0.35.0
whitenoise==6.9.0
